refactor: split `numpy` into `array_module`
