+++ src/awkward/_nplikes/__init__.py
@@ -1,0 +1,5 @@
+import awkward._nplikes.dtypes  # noqa: F401
+from awkward._nplikes.cupy import Cupy  # noqa: F401
+from awkward._nplikes.numpy import Numpy  # noqa: F401
+from awkward._nplikes.numpylike import NumpyLike  # noqa: F401
+from awkward._nplikes.typetracer import TypeTracer  # noqa: F401
